+++ modules/contracts/src.sol/CMCAdjudicator.sol
@@ -2,6 +2,7 @@
 pragma solidity ^0.7.1;
 pragma experimental ABIEncoderV2;
 
+import "./interfaces/Commitment.sol";
 import "./interfaces/ICMCAdjudicator.sol";
 import "./interfaces/ITransferDefinition.sol";
 import "./interfaces/Types.sol";
@@ -30,10 +31,7 @@ contract CMCAdjudicator is CMCCore, CMCAccountant, ICMCAdjudicator {
   }
 
   modifier validateTransfer(CoreTransferState calldata cts) {
-    require(
-      cts.channelAddress == address(this),
-      "CMCAdjudicator: INVALID_TRANSFER"
-    );
+    require(cts.channelAddress == address(this), "CMCAdjudicator: INVALID_TRANSFER");
     _;
   }
 
@@ -91,10 +89,7 @@ contract CMCAdjudicator is CMCCore, CMCAccountant, ICMCAdjudicator {
     validateChannel(ccs)
   {
     // Verify that the given channel state matches the stored one
-    require(
-      hashChannelState(ccs) == channelDispute.channelStateHash,
-      "CMCAdjudicator: INVALID_CHANNEL_HASH"
-    );
+    require(hashChannelState(ccs) == channelDispute.channelStateHash, "CMCAdjudicator: INVALID_CHANNEL_HASH");
 
     // We need to be in defund phase for that
     require(inDefundPhase(), "CMCAdjudicator: INVALID_PHASE");
@@ -178,10 +173,7 @@ contract CMCAdjudicator is CMCCore, CMCAccountant, ICMCAdjudicator {
     require(transferDispute.transferDisputeExpiry != 0, "CMCAdjudicator: TRANSFER_NOT_DISPUTED");
 
     // Verify that the given transfer state matches the stored one
-    require(
-      hashTransferState(cts) == transferDispute.transferStateHash,
-      "CMCAdjudicator: INVALID_TRANSFER_HASH"
-    );
+    require(hashTransferState(cts) == transferDispute.transferStateHash, "CMCAdjudicator: INVALID_TRANSFER_HASH");
 
     // We can't defund twice
     require(!transferDispute.isDefunded, "CMCAdjudicator: TRANSFER_ALREADY_DEFUNDED");
@@ -192,10 +184,7 @@ contract CMCAdjudicator is CMCCore, CMCAccountant, ICMCAdjudicator {
     if (block.number < transferDispute.transferDisputeExpiry) {
       // Before dispute expiry, responder can resolve
       require(msg.sender == cts.responder, "CMCAdjudicator: INVALID_MSG_SENDER");
-      require(
-        keccak256(encodedInitialTransferState) == cts.initialStateHash,
-        "CMCAdjudicator: INVALID_TRANSFER_HASH"
-      );
+      require(keccak256(encodedInitialTransferState) == cts.initialStateHash, "CMCAdjudicator: INVALID_TRANSFER_HASH");
       ITransferDefinition transferDefinition = ITransferDefinition(cts.transferDefinition);
       balance = transferDefinition.resolve(
         abi.encode(cts.balance),
@@ -233,15 +222,14 @@ contract CMCAdjudicator is CMCCore, CMCAccountant, ICMCAdjudicator {
     return _getTotalDepositsBob(assetId);
   }
 
-  // TODO: include commitment type
   function verifySignatures(
     CoreChannelState calldata ccs,
     bytes calldata aliceSignature,
     bytes calldata bobSignature
   ) internal pure {
-    bytes32 ccsHash = hashChannelState(ccs);
-    require(ccsHash.checkSignature(aliceSignature, ccs.alice), "CMCAdjudicator: INVALID_ALICE_SIG");
-    require(ccsHash.checkSignature(bobSignature, ccs.bob), "CMCAdjudicator: INVALID_BOB_SIG");
+    bytes32 commitment = keccak256(abi.encode(CommitmentType.ChannelState, hashChannelState(ccs)));
+    require(commitment.checkSignature(aliceSignature, ccs.alice), "CMCAdjudicator: INVALID_ALICE_SIG");
+    require(commitment.checkSignature(bobSignature, ccs.bob), "CMCAdjudicator: INVALID_BOB_SIG");
   }
 
   function verifyMerkleProof(
@@ -261,7 +249,6 @@ contract CMCAdjudicator is CMCCore, CMCAccountant, ICMCAdjudicator {
   }
 
   function hashChannelState(CoreChannelState calldata ccs) internal pure returns (bytes32) {
-    // TODO: include commitment type
     return keccak256(abi.encode(ccs));
   }