+++ modules/contracts/src.sol/interfaces/Commitment.sol
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: UNLICENSED
+pragma solidity ^0.7.1;
+pragma experimental ABIEncoderV2;
+
+enum CommitmentType {
+  ChannelState,
+  WithdrawData
+}