+++ modules/documentation/docs/changelog.md
@@ -4,6 +4,12 @@
 
 ## @connext/{types,utils,contracts,protocol,engine,browser-node}@0.0.9, @connext/{server-node}@0.0.10
 
+- introduce and use commitment types
+- reconcile hashing
+- remove commitment type
+
+## @connext/{types,utils,contracts,protocol,engine,browser-node}@0.0.9, @connext/{server-node}@0.0.10
+
 - implement generalized withdrawals
 - fix and minor improvements around asset transfers
 - move dispute methods into engine